fix/doc: cleaning code web-client/app(2)
