+++ web-client/src/components/common/Avatar.tsx
