+++ web-client/src/components/index.ts
@@ -1,0 +1,38 @@
+/* Work in progress */
+
+// Common components
+// export { default as ActionButton } from "./common/ActionButton";
+// export { default as Avatar } from "./common/Avatar";
+// export { default as Button } from "./common/Button";
+// export { default as Card } from "./common/Card";
+// export { default as DropdownMenu } from "./common/DropdownMenu";
+// export { default as Label } from "./common/Label";
+// export { default as SingleTag } from "./common/SingleTag";
+// export { default as ToggleSwitch } from "./common/ToggleSwitch";
+
+// // Forms
+// export { default as Input } from "./forms/Input";
+// export { default as ListInputWrapper } from "./forms/ListInputWrapper";
+// export { default as TagInput } from "./forms/TagInput";
+
+// // Layout
+// export { default as Header } from "./layout/Header";
+// export { default as LandingTheme } from "./layout/LandingTheme";
+
+// // Lists
+// export { default as ItemCard } from "./list/ItemCard";
+// export { default as ListCard } from "./list/ListCard";
+// export { default as ListCardContainer } from "./list/ListCardContainer";
+
+// // Map
+// export { default as Map } from "./map/Map";
+
+// // Pages
+// export { default as SetLocation } from "./pages/SetLocation";
+// export { default as SupermarketCard } from "./pages/SupermarketCard";
+
+// // Products
+// export { default as ProductGrid } from "./products/ProductGrid";
+// export { default as ProductItem } from "./products/ProductItem";
+// export { default as ProductList } from "./products/ProductList";
+// export { default as ProductTags } from "./products/ProductTags";
