+++ web-client/src/components/utils/foodEmojis.ts
